% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/behave_pupil.R
\name{behave_pupil}
\alias{behave_pupil}
\title{culls  important behavioral data from the sample report}
\usage{
behave_pupil(datafile, omiterrors = TRUE, behave_colnames = NULL)
}
\arguments{
\item{datafile}{raw pupil data}

\item{omiterrors}{if ture returns RTs without errors for RT analysis or dataset with all error data for error analysis}

\item{behave_colnames}{specify variables of interest}
}
\value{
data frame with or withour errors
}
\description{
culls  important behavioral data from the sample report
}
