% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blink_detect.R
\name{blink_detect}
\alias{blink_detect}
\title{Take pupil file and detect blinks
merge blinks with sample report}
\usage{
blink_detect(samp)
}
\arguments{
\item{samp}{pupil data file}
}
\value{
merged data file with blinks detected
}
\description{
Take pupil file and detect blinks
merge blinks with sample report
}
