% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blink_velocity.R
\name{blink_velocity}
\alias{blink_velocity}
\title{This function takes a dataframe and applies a simple velocity algorithm to detect blinks in
the data based off velocity of pupil size changes. The algorithm has four
parameters that need to be adjusted based on the
particulars of the data: the amount of smoothing, the (negative) onset velocity threshold,
the (positive) reversal velocity threshold, and the temporal margin}
\usage{
blink_velocity(
  df,
  neg_velocity_thresh = -5,
  pos_velocity_thresh = 5,
  fillback = 10,
  fillforward = 10,
  hz = 250
)
}
\arguments{
\item{df}{data frame.}

\item{neg_velocity_thresh}{threshold for onset of blinks neg number here we use default of -5}

\item{fillback}{extend blinks backward}

\item{fillforward}{extend blinks forward}

\item{hz}{recording frequency of ET}

\item{pos_velocity_thres}{threshold for offset of blinks neg number here we use default of 5}
}
\value{
data frame containing pupil values where blinks have been linear interpolated.
}
\description{
This function takes a dataframe and applies a simple velocity algorithm to detect blinks in
the data based off velocity of pupil size changes. The algorithm has four
parameters that need to be adjusted based on the
particulars of the data: the amount of smoothing, the (negative) onset velocity threshold,
the (positive) reversal velocity threshold, and the temporal margin
}
