% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_mad.R
\name{calc_mad}
\alias{calc_mad}
\title{calculates the median absoulte deviation (MAD) and sets the threshold to filter on.}
\usage{
calc_mad(max_dilation, n = 16)
}
\arguments{
\item{max_dilation}{pupil file returned from speed_pupil}

\item{n}{constant used}
}
\value{
threhold value to filter on
}
\description{
calculates the median absoulte deviation (MAD) and sets the threshold to filter on.
}
