% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code_poly.R
\name{code_poly}
\alias{code_poly}
\title{Build polynomial predictor variables}
\usage{
code_poly(
  df = NULL,
  predictor = NULL,
  poly.order = NULL,
  orthogonal = TRUE,
  draw.poly = TRUE
)
}
\arguments{
\item{df}{data frame, should not contain any variables called "predictor"}

\item{predictor}{string name of predictor variable}

\item{poly.order}{integer order of polynomial to be created}

\item{orthogonal}{logical value indicating whether polynomial should be
orthogonal (default) or natural (aka raw)}

\item{draw.poly}{logical value indicating whether to create a graph showing
tranformed polynomial predictor values, defaults to TRUE}
}
\value{
Returns a data frame containing the original data and at least two
new columns: "predictor".Index, and a column for each order of the
polynomial-transformed predictor
}
\description{
Takes a data frame, name of predictor variable, and polynomial order. Creates
polynomial-transformed predictor variables, adds them to the data frame and
returns the result. Original data frame is unchanged, remember to assign the
result.
}
\examples{
WordLearnEx.gca <- code_poly(df=WordLearnEx, predictor="Block", poly.order=2)
Az.gca <- code_poly(df=Az, predictor="Time", poly.order=2, orthogonal=FALSE)
}
