% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_groups.R
\name{compare_groups}
\alias{compare_groups}
\title{Compare groups on a set of variables.}
\usage{
compare_groups(dat, Group)
}
\arguments{
\item{dat}{data frame}

\item{Group}{string name of grouping variable}
}
\value{
Returns a data frame with rows for each numeric variable in the input data frame (plus N, for number of observations) and columns for mean, SD, t-value, and p-value.
}
\description{
For each numeric variable in a data frame, computes number of observations, mean, and standard deviation, for each group, and t- and p-values for the t-test comparing the groups.
}
\examples{
compare_groups(SND, "SemNear_Cond")
}
