% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_monocular_mean.R
\name{compute_monocular_mean}
\alias{compute_monocular_mean}
\title{Compute a Monocular Mean Pupil Value}
\usage{
compute_monocular_mean(x1, x2)
}
\arguments{
\item{x1}{Numeric vector for the left pupil.}

\item{x2}{Numeric vector for the right pupil.}
}
\value{
A numeric vector of monocular mean pupil values.
}
\description{
A "monocular mean" averages both eyes together. If data is available in just
one eye, use the available value as the mean. If both eyes are missing, the
result is \code{NA}.
}
