% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Crawford_Howell.R
\name{Crawford_Howell}
\alias{Crawford_Howell}
\title{Crawford-Howell (1998) t-test for case-control comparisons.}
\usage{
Crawford_Howell(case, control)
}
\arguments{
\item{case}{single case score}

\item{control}{vector of control scores}
}
\value{
Returns a data frame with three values: t, df, and p.
}
\description{
Computes Crawford-Howell t-test comparing a single case to a set of controls.
}
\examples{
Crawford_Howell(649, c(243, 175, 231, 164, 236, 131, 302, 303, 363, 287, 205, 297, 144, 789, 210, 245))
}
