% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cursive.R, R/data_documentation.R
\docType{data}
\name{cursive}
\alias{cursive}
\title{Judgments to cursive vs. type-print words in LDT}
\format{A data frame with 341 rows and 4 variables
\describe{
 \item{subject}{subject id, N=9}
 \item{script}{Cursive or Print}
 \item{timebinonset}{trial onset}
 \item{baselinep}{baseline corrected pupil size in mm}
 
}}
\usage{
cursive

cursive
}
\description{
Aggregate baseline corrected pupillary data for a LDT task that required participants to judge the lexical status of cursive and type-print stimuli.

Aggregate baseline corrected pupillary data for a LDT task that required participants to judge the lexical status of cursive and type-print stimuli.
}
\keyword{datasets}
