% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{cursive_new}
\alias{cursive_new}
\title{Judgments to cursive vs. type-print words in LDT}
\format{A data frame with 1312 rows and 4 variables
\describe{
 \item{subject}{subject id, N=41}
 \item{script}{Cursive or Print}
 \item{timebinonset}{trial onset}
 \item{baselinep}{baseline corrected pupil size}

}}
\usage{
cursive_new
}
\description{
Aggregate baseline corrected pupillary data for a LDT task that required participants to judge the lexical status of cursive and type-print stimuli.
}
\keyword{datasets}
