% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downsample.pupil.R
\name{downsample.pupil}
\alias{downsample.pupil}
\title{Downsample pupil data}
\usage{
downsample.pupil(df, bin.length = NULL)
}
\arguments{
\item{df}{dataframe}

\item{bin.length}{Length of bins to average}
}
\description{
This function will reduce the sampling fequency
}
