% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynamic_pupil.R
\name{dynamic_pupil}
\alias{dynamic_pupil}
\title{calc dynamic pupil range (Ayasse et al., 2017; Piquado et al., 2010))
must have min pupil and max pupil size merged into df to work}
\usage{
dynamic_pupil(pupil_col, dmin = NA, dmax = NA)
}
\arguments{
\item{pupil_col}{pupil col}

\item{dmin}{mean pupil size during the ‘bright’ condition (i.e., minimum pupil size)}

\item{dmax}{mean pupil size during the ‘dark’ condition (i.e., max pupil size)}
}
\value{
return dynamic pupil size scaled to P
}
\description{
calc dynamic pupil range (Ayasse et al., 2017; Piquado et al., 2010))
must have min pupil and max pupil size merged into df to work
}
