% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extend_blinks.R
\name{extend_blinks}
\alias{extend_blinks}
\title{Extends blinks if not already done in SR Data Viewer}
\usage{
extend_blinks(values, fillback = 5, fillforward = 5)
}
\arguments{
\item{values}{column with pupil values}

\item{fillback}{NA extend backward}

\item{fillforward}{NA extend forward}
}
\value{
vector containing extended blinks
}
\description{
Extends blinks if not already done in SR Data Viewer
}
