% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendblinks.R
\name{extendblinks}
\alias{extendblinks}
\title{Extends blinks if not already done in SR Data Viewer}
\usage{
extendblinks(values, fillback = 5, fillforward = 5)
}
\arguments{
\item{values}{column with pupil values}

\item{fillback}{NA extend backward}

\item{fillforward}{NA extend forward}
}
\value{
vector containing extended blinks
}
\description{
Extends blinks if not already done in SR Data Viewer
}
