% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{FunctThemePts}
\alias{FunctThemePts}
\title{Fixation of semantically related objects during spoken word comprehension by participants with left hemisphere stroke}
\format{
A data frame with 8199 rows and 8 variables:
\describe{
\item{Subject}{subject ID, factor}
\item{Condition}{semantic relatedness condition: Function (e.g., broom - sponge) or Thematic (e.g., broom - dustpan)}
\item{Object}{object type: Target, semantically related Competitor, or Unrelated}

\item{Time}{time (in ms) relative to word onset in 50ms bins; negative numbers indicate preview period (visual display before word onset)}
\item{meanFix}{fixation proportion (subject mean): proportion of correct-response trials on which the object was fixated}
\item{sumFix}{fixation sum: number of correct-response trials on which the object was fixated (i.e., numerator for meanFix)}
\item{N}{number of correct-response trials (i.e., denominator for meanFix)}
}
}
\source{
Kalenine, Mirman, & Buxbaum (2012, Front. Hum. Neurosci., 6:106, 1-12) \url{http://journal.frontiersin.org/Journal/10.3389/fnhum.2012.00106}
}
\usage{
FunctThemePts
}
\description{
Fixation time course data for participants with left hemisphere stroke from a spoken word-to-picture matching study with semantic distractors in the display.
}
\keyword{datasets}
