% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gazer.R
\name{gazeClusters}
\alias{gazeClusters}
\title{Application of the the DBSCAN clustering algorithm for eye tracking data.}
\usage{
gazeClusters(dat, eps, minPts = 5, recRate = NULL)
}
\arguments{
\item{dat}{A data frame or a matrix with three columns named: X, Y, time.}

\item{eps}{Size (radius) of the epsilon neighborhood.}

\item{minPts}{Number of minimum points required in the eps neighborhood for core points (including the point itself). The default value is 5 points.}

\item{recRate}{Time in ms between each eye-tracking recording. For instance 1 means 1 ms between two observations or a recording frequency of 1000 Hz.}
}
\value{
Provides a numeric vector indicating wheather an eye-tracking recording belongs to an eye-movement or to a fixation.
zero 0 means that the recording belongs to an eye-movement. Any other positive number indicates a fixation ID.
In addition, this function also returns a vector of the fixations IDs and their durations.
}
\description{
Application of the the DBSCAN clustering algorithm for eye tracking data.
}
