\name{gazer-package}
\alias{gazer-package}
\alias{gazer}
\docType{package}
\title{
\packageTitle{gazer}
}
\description{
\packageDescription{gazer}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{gazer}
\packageIndices{gazer}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{gazer}

Maintainer: \packageMaintainer{gazer}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R ~~
~~ documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
