% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_edf2asc_bat.R
\name{generate_edf2asc_bat}
\alias{generate_edf2asc_bat}
\title{Takes edf files and converts them to asc
This code was written by Dr. Holger Mitterer}
\usage{
generate_edf2asc_bat(homeDir, subDirs, deleteExisting = F)
}
\arguments{
\item{homeDir}{dir of files}

\item{subDirs}{end location of files}

\item{deleteExisting}{overwrite file - set to false}
}
\description{
Takes edf files and converts them to asc
This code was written by Dr. Holger Mitterer
}
