% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_gaze_diagnostics.R
\name{get_gaze_diagnostics}
\alias{get_gaze_diagnostics}
\title{Calculate gaze track diagnostics}
\usage{
get_gaze_diagnostics(gaze, screen_size = c(1024, 768))
}
\arguments{
\item{gaze}{fixation list from fixation report}

\item{screen_size}{size of the screen in pixels. Defaults to c(1024, 768)}
}
\value{
data frame containing gaze diagnostics
}
\description{
Use fixation data to calculate two diagnostics of gaze track quality: non-fixation time and out-of-bounds proportion.
}
