% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ranef.R
\name{get_ranef}
\alias{get_ranef}
\title{Get random effects for individual effect size quantification}
\usage{
get_ranef(model = NULL, ranef_name = NULL)
}
\arguments{
\item{model}{(g)lmerMod object}

\item{ranef_name}{string name of random effect (right of pipe in (g)lmer call)}
}
\value{
Returns a data frame containing the random effect estimates
}
\description{
Takes a model and name of a random effect. Returns a data frame containing those random effect estimates arranged and ready for quantifying individual effect sizes.
}
\examples{
re <- get_ranef(m.FT, "Subject:Object:Condition") #using FunctTheme example model
}
