% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_trackloss.R
\name{get_trackloss}
\alias{get_trackloss}
\title{Calculate trackloss and remove trials and subjects with excessive trackloss}
\usage{
get_trackloss(gaze, screen_size = c(1024, 768), missingthresh = 0.2)
}
\arguments{
\item{gaze}{fixation list from fixation report}

\item{screen_size}{size of the screen in pixels. Defaults to c(1024, 768)}

\item{missingthresh}{threshold to throw out trials and subjs}
}
\value{
data frame containing gaze df with trackloss by subjs and trial. Trials and Subjs with excessive trackloss removed
}
\description{
Calculate trackloss and remove trials and subjects with excessive trackloss
}
