% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hanning_filter.R
\name{hanning_filter}
\alias{hanning_filter}
\title{Perfroms hanning smoothing on pupil data}
\usage{
hanning_filter(pupil, degree = 11)
}
\arguments{
\item{pupil}{column with pupil values}

\item{degree}{default of 11.}
}
\value{
vector containing hanning filtered pupil data
}
\description{
Perfroms hanning smoothing on pupil data
}
