% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/legend_positioning.R
\name{legend_positioning}
\alias{legend_positioning}
\title{Set legend position}
\usage{
legend_positioning(pos, edge_color = "black", legend_fill = "white")
}
\arguments{
\item{pos}{Legend position as a two-element vector}

\item{edge_color}{Legend edge color, defaults to black}

\item{legend_fill}{Legend fill color, defaults to white}
}
\description{
A wrapper function to make setting legend position easier when using ggplot.
}
\examples{
last_plot() + legend_positioning(c(0,1)) #put legend in upper left corner
last_plot() + legend_positioning(c(1,0)) #put legend in lower right corner
}
