% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_gazer.R
\name{make_gazer}
\alias{make_gazer}
\title{Prepare data for pre-processing in Gazer}
\usage{
make_gazer(
  data,
  subject = "subject",
  trial = "trial",
  time = "time",
  x = "x",
  y = "y",
  pupil = NULL
)
}
\arguments{
\item{data}{a raw, long form dataframe organised by subject, trial, and time.}

\item{subject}{column name indicating subject ID}

\item{trial}{column name indicating trial ID. This should be unique for participants}

\item{time}{column name indicating time column (should be numeric)}

\item{x}{gaze x coordinates}

\item{y}{gaze y coordinates}

\item{pupil}{name of pupil column}
}
\value{
A dataframe ready to use in gazer
}
\description{
This will make sure your data is in the right format for processing.
This package is designed to deal with data at it comes out of the eyetracker
in a long-form csv style format. Thus data input here would be a long
dataframe, wherein each row is a single frame collected by the eyetracker.
}
