% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean_max_pupil.R
\name{mean_max_pupil}
\alias{mean_max_pupil}
\title{Calculates mean and max pupil size for each trial}
\usage{
mean_max_pupil(datafile)
}
\arguments{
\item{datafile}{raw pupil data}
}
\value{
data frame containing data with mean and max pupil size appended
}
\description{
Calculates mean and max pupil size for each trial
}
