% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_gazer_files.R
\name{merge_gazer_files}
\alias{merge_gazer_files}
\title{Merge eye files from a ufolder
if filetype is from sr renames column variables and makes them lower case
Places time in ms
if edf files just reads them in}
\usage{
merge_gazer_files(
  file_list,
  blink_colname = "blink",
  pupil_colname = "pupil",
  filetype = "sr"
)
}
\arguments{
\item{file_list}{path to .xls files}

\item{blink_colname}{name of your blink colname:AVERAGE_IN_BLINK, LEFT_, #RIGHT}

\item{pupil_colname}{name of your pupil colname:AVERAGE_IN_Pupil, LEFT_, #RIGHT}

\item{filetype}{if reading raw edf file use 'edf' if using sr files use type 'sr'}
}
\value{
data frame containing pupil data
}
\description{
Merge eye files from a ufolder
if filetype is from sr renames column variables and makes them lower case
Places time in ms
if edf files just reads them in
}
