% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_pupil.R
\name{merge_pupil_files}
\alias{merge_pupil_files}
\title{Merge pupil files from a folder}
\usage{
merge_pupil_files(file_list)
}
\arguments{
\item{file_list}{path to .xls files}
}
\value{
data frame containing pupil data
}
\description{
Renames column variables and makes them lower case
Places time in ms
}
