% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_tobii.R
\name{merge_tobii}
\alias{merge_tobii}
\title{Merges files taken from a X2-30 Tobii tracker created from a PsychoPy experiment using the Python package
Titta which interacts with the Tobii SDK.
Removes calibration and validation messages from main file
Puts time in milliseconds
Averages left and right pupil data
Creates a trial col --note you need to make sure your recording has trial messages
Changes colnames so it is ready to be used in gazeR}
\usage{
merge_tobii(
  file_list,
  part_colname = "PID",
  behave_colnames = "cond",
  trial_msg = "startfix_",
  type = "X230"
)
}
\arguments{
\item{file_list}{a vector of file names}

\item{part_colname}{name of your participant id}

\item{behave_colnames}{name of your behave colnames}

\item{trial_msg}{name of message used to get trials. In this case at start of fix I sent msg "startfix_trialnum"}

\item{type}{right now work with X230}
}
\value{
DF
}
\description{
Merges files taken from a X2-30 Tobii tracker created from a PsychoPy experiment using the Python package
Titta which interacts with the Tobii SDK.
Removes calibration and validation messages from main file
Puts time in milliseconds
Averages left and right pupil data
Creates a trial col --note you need to make sure your recording has trial messages
Changes colnames so it is ready to be used in gazeR
}
