% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missing_pupil_count.R
\name{missing_pupil_count}
\alias{missing_pupil_count}
\title{Calculate percent missing data by subject and by trial}
\usage{
missing_pupil_count(datafile, missingthresh = 0.3)
}
\arguments{
\item{missingthresh}{user-specified threshold for missing data.}
}
\value{
data frame containing missing data information
}
\description{
Calculate percent missing data by subject and by trial
}
