% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moving_average_pupil.R
\name{moving_average_pupil}
\alias{moving_average_pupil}
\title{Moving average function}
\usage{
moving_average_pupil(x, n = 5, centered = TRUE)
}
\arguments{
\item{x}{data}

\item{n}{sed for moving window}

\item{centered}{whether moving average window should be centered (TRUE) or trailing (FALSE)}
}
\value{
return sum divided by count
}
\description{
Creates a moving average using a specified averageing window
if n is 5, then it uses the target value, 2 preceding and 2 following
}
