% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/movingaverage.R
\name{movingaverage}
\alias{movingaverage}
\title{Moving average function
Creates a moving average using a specified averageing window
if n is 5, then it uses the target value,
 2 preceding and 2 following
 @param x data 
 @param  n #sed for moving window
 @param centered 
 @return  return sum divided by count
 @export}
\usage{
movingaverage(x, n = 5, centered = TRUE)
}
\description{
Moving average function
Creates a moving average using a specified averageing window
if n is 5, then it uses the target value,
 2 preceding and 2 following
 @param x data 
 @param  n #sed for moving window
 @param centered 
 @return  return sum divided by count
 @export
}
