% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/noise_based_blink_detection.R
\name{noise_based_blink_detection}
\alias{noise_based_blink_detection}
\title{Extends blinks if not already done in SR Data Viewer}
\usage{
noise_based_blink_detection(pupil_data, hz)
}
\arguments{
\item{pupil_data}{column with pupil values}

\item{hz}{= sampling rate of eye tracker}
}
\value{
blinks
}
\description{
Extends blinks if not already done in SR Data Viewer
}
