% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/onset_pupil.R
\name{onset_pupil}
\alias{onset_pupil}
\title{Zero Time Index to Event Onset}
\usage{
onset_pupil(time, marker, event)
}
\arguments{
\item{time}{Numeric vector of sample indices or time stamps corresponding to
a single time series.}

\item{marker}{Vector containing markers of trial events.}

\item{event}{Label in \code{marker} that indicates the onset event to which
the time series should be zeroed. If this label is a character string, it
must be enclosed in quotation marks.}
}
\value{
Numeric vector that zeroes \code{time} to the trial onset. In the
event that no onset marker is found, a warning is issued and a vector of
\code{NA}s is returned.
}
\description{
Centers the sample index or time stamp of your time series to a reference
marker, such that 0 always coincides with the onset of a particular event.
}
