% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_edf_messages.R
\name{parse_edf_messages}
\alias{parse_edf_messages}
\title{Take EDF files and return message events to extract behav data
creates pupil column that is ambigious to whether you sampled from left eye, right eye, or both (takes the average)}
\usage{
parse_edf_messages(file_list, output.dir)
}
\arguments{
\item{file_list}{directory to edf files}

\item{output.dir}{directory to save new cleaned files}
}
\description{
Take EDF files and return message events to extract behav data
creates pupil column that is ambigious to whether you sampled from left eye, right eye, or both (takes the average)
}
