% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_fixation_report.R
\name{read_fixation_report}
\alias{read_fixation_report}
\title{Read EyeDataViewer fixation report}
\usage{
read_fixation_report(
  filename,
  screen_size = c(1024, 768),
  name_changes = c(Subject = "RECORDING_SESSION_LABEL", ACC = "StimSlide.ACC", RT =
    "StimSlide.RT", TargetLoc = "CorrectPort"),
  plot_fix_scatter = TRUE,
  pages = 1
)
}
\arguments{
\item{filename}{string containing full filename of fixation report}

\item{screen_size}{size of the screen in pixels. Used to compute track quality diagnostics. Defaults to c(1024, 768)}

\item{name_changes}{set of variable name pairs for improving variable names. Should be formatted as named character vector with new names as values and old names as names; see plyr::rename()}

\item{plot_fix_scatter}{logical indicating whether a scatterplot of fixations should be created}

\item{page}{outputs a plots 4 x 3 and displays the first page. Change number to see other pages if you have more Ps}
}
\value{
gaze data frame containing fixation report
}
\description{
Reads in a fixation report created by EyeLink's EyeDataViewer application
}
