% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/readFixationReport.R
\name{readFixationReport}
\alias{readFixationReport}
\title{Read EyeDataViewer fixation report}
\usage{
readFixationReport(filename, screen_size = c(1024, 768),
  name_changes = c(RECORDING_SESSION_LABEL = "Subject", StimSlide.ACC = "ACC",
  StimSlide.RT = "RT", CorrectPort = "TargetLoc"), plot_fix_scatter = TRUE)
}
\arguments{
\item{filename}{string containing full filename of fixation report}

\item{screen_size}{size of the screen in pixels. Used to compute track quality diagnostics. Defaults to c(1024, 768)}

\item{name_changes}{set of variable name pairs for improving variable names. Should be formatted as named character vector with new names as values and old names as names; see plyr::rename()}

\item{plot_fix_scatter}{logical indicating whether a scatterplot of fixations should be created}
}
\value{
gaze data frame containing fixation report
}
\description{
Reads in a fixation report created by EyeLink's EyeDataViewer application
}

