% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smooth_interpolate_pupil.R
\name{smooth_interpolate_pupil}
\alias{smooth_interpolate_pupil}
\title{Smooths the pupil trace before linear or cubic interpolation.}
\usage{
smooth_interpolate_pupil(
  datafile,
  pupil = "pupil",
  extendpupil = "extendpupil",
  extendblinks = FALSE,
  step.first = "smooth",
  filter = "moving",
  maxgap = Inf,
  type = NULL,
  hz = NA,
  n = NA
)
}
\arguments{
\item{datafile}{data frame.}

\item{pupil}{name of pupil column}

\item{extendpupil}{name of pupil col that was extended}

\item{extendblinks}{blinks already extended  in data frame}

\item{step.first}{whether you want to smooth/interpolate or interpolate/smooth}

\item{filter}{use either moving and hann filter}

\item{maxgap}{max number of NAs to interpolate.}

\item{type}{type of interpolation (linear or cubic)}

\item{hz}{recording frequency of ET}

\item{n}{moving average window}
}
\value{
df containing interpolated data
}
\description{
Smooths the pupil trace before linear or cubic interpolation.
}
