% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/speed_pupil.R
\name{speed_pupil}
\alias{speed_pupil}
\title{Calculates max absoulte dilation speed as suggested in Kret & Sjak-Shie (in press)
calculates the median absoulte deviation (MAD) and sets the threshold to filter on.}
\usage{
speed_pupil(pup, time)
}
\arguments{
\item{pup}{column listing pupil sizes}

\item{time}{column in data file listing time}
}
\value{
data with pupil speeds for each time point
}
\description{
Calculates max absoulte dilation speed as suggested in Kret & Sjak-Shie (in press)
calculates the median absoulte deviation (MAD) and sets the threshold to filter on.
}
