% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baseline_correction_pupil.R
\name{baseline_correction_pupil}
\alias{baseline_correction_pupil}
\title{Performs linear (substraction) baseline correction relative to desired stim_onset and off_set}
\usage{
baseline_correction_pupil(datafile, pupil_colnames = NULL,
  baseline_window = NA)
}
\arguments{
\item{datafile}{raw pupil data}

\item{pupil_colnames}{name of your pupil colname you want baseline corrected}

\item{baseline_window}{user-specified threshold for baseline window.}
}
\value{
data frame containing baseline corrected data from event of interest
}
\description{
Performs linear (substraction) baseline correction relative to desired stim_onset and off_set
}
