% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baseline_correction_pupil.R,
%   R/baseline_correction_pupil_mean.R
\name{baseline_correction_pupil}
\alias{baseline_correction_pupil}
\title{Performs linear (sub) or divisive baseline correction relative to desired stim_onset and off_set}
\usage{
baseline_correction_pupil(datafile, pupil_colname = NULL,
  baseline_window = NA, baseline_method = "sub")

baseline_correction_pupil(datafile, pupil_colname = NULL,
  baseline_window = NA, baseline_method = "sub")
}
\arguments{
\item{datafile}{raw pupil data}

\item{pupil_colname}{name of your pupil colname you want baseline corrected}

\item{baseline_window}{user-specified threshold for baseline window.}

\item{baseline}{correction method. Default is sub but can also include divisive}

\item{datafile}{raw pupil data}

\item{pupil_colname}{name of your pupil colname you want baseline corrected}

\item{baseline_window}{user-specified threshold for baseline window.}

\item{baseline}{correction method. Default is sub but can also include divisive}
}
\value{
data frame containing baseline corrected data from event of interest

data frame containing baseline corrected data from event of interest
}
\description{
Performs linear (sub) or divisive baseline correction relative to desired stim_onset and off_set

Performs linear (sub) or divisive baseline correction relative to desired stim_onset and off_set
}
