% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baseline_correction_pupil_msg.R
\name{baseline_correction_pupil_msg}
\alias{baseline_correction_pupil_msg}
\title{Performs linear (sub) or divisive baseline correction relative to desired stim_onset and off_set with messages.}
\usage{
baseline_correction_pupil_msg(datafile, pupil_colname = NULL,
  baseline_dur = 500, event_offset = "event_offset",
  baseline_method = "sub")
}
\arguments{
\item{datafile}{raw pupil data}

\item{pupil_colname}{name of your pupil colname you want baseline corrected}

\item{baseline_dur}{duration of baseline period from stim offset}

\item{event_offset}{message name for stim onset}

\item{baseline_method}{correction method. Default is sub but can also include divisive}
}
\value{
data frame containing baseline corrected data from event of interest
}
\description{
Performs linear (sub) or divisive baseline correction relative to desired stim_onset and off_set with messages.
}
