% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binify_fixations.R
\name{binify_fixations}
\alias{binify_fixations}
\title{Convert fixation list into time bins}
\usage{
binify_fixations(gaze, binSize = 20, keepCols = c("Subject",
  "TrialNumber", "Target", "T"), maxTime = NULL)
}
\arguments{
\item{gaze}{a data frame of fixations, as returned by readFixationReport}

\item{binSize}{length of bins, in ms}

\item{keepCols}{set of columns to keep, fixation start and end times will be dropped}

\item{maxTime}{optional parameter to set maximum trial length}
}
\value{
data frame containing gaze data arranged by bin
}
\description{
Takes a data frame list of fixations and returns a time series
with the fixations expanded into time bins of specified duration.
}
