% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_missing_pupil.R
\name{count_missing_pupil}
\alias{count_missing_pupil}
\title{Calculate percent missing data by subject and by trial}
\usage{
count_missing_pupil(datafile, missingthresh = 0.2)
}
\arguments{
\item{missingthresh}{user-specified threshold for missing data. set to .2}
}
\value{
data frame containing missing data information
}
\description{
Calculate percent missing data by subject and by trial
}
