% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{cursive_agg_data}
\alias{cursive_agg_data}
\title{Judgments to cursive vs. type-print words in LDT}
\format{A data frame with 52 rows and 3 variables
\describe{
\item{script}{Cursive or Print}
\item{timebins}{trial onset}
\item{aggbaseline}{baseline corrected pupil size}

}}
\usage{
cursive_agg_data
}
\description{
Aggregate baseline corrected pupillary data for a LDT task that required participants to judge the lexical status of cursive and type-print stimuli.
}
\keyword{datasets}
