% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downsample_gaze.R
\name{downsample_gaze}
\alias{downsample_gaze}
\title{Downsample gaze data}
\usage{
downsample_gaze(dataframe, bin.length = 50, timevar = "time",
  aggvars = list("subject", "condition", "target", "trial", "object",
  "timebin"))
}
\arguments{
\item{dataframe}{dataframe}

\item{bin.length}{Length of time bins}

\item{aggvars}{list of variable names to group by for aggregation}
}
\description{
This function will combine gaze samples into time bins
}
