% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downsample_pupil.R
\name{downsample_pupil}
\alias{downsample_pupil}
\title{Downsample pupil data}
\usage{
downsample_pupil(dataframe, bin.length = NULL)
}
\arguments{
\item{dataframe}{dataframe}

\item{bin.length}{Length of bins to average}
}
\description{
This function will reduce the sampling fequency
}
