% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_messages_asc.R
\name{find_messages_asc}
\alias{find_messages_asc}
\title{Take asc files and return important events
Collect trial variables for your experiment.
#' This code was written by Dr. Holger Mitterer}
\usage{
find_messages_asc(
  dirList,
  homeDir = "./",
  time2extract = NULL,
  msg2extract = NULL,
  vars2extract = NULL,
  nOfTrials = -1
)
}
\arguments{
\item{dirList}{dir of edf files}

\item{homeDir}{dir of files}

\item{time2extract}{a list of strings that match message texts for which only the time should be retained.}

\item{vars2extract}{A list of Trial variables used by Experiment Builder that should be extracted value only, with column names set by the variable name}

\item{msg2extractA}{list of messages to extract fully based on the first word of that message output will be string variable}
}
\description{
Take asc files and return important events
Collect trial variables for your experiment.
#' This code was written by Dr. Holger Mitterer
}
