% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_messages.R, R/find_messages_edf.R
\name{find_messages_edf}
\alias{find_messages_edf}
\title{Take EDF files and return a tiblle with important behavioral events}
\usage{
find_messages_edf(file_list, varnames, patterns, output_directory)

find_messages_edf(file_list, varnames, patterns, output_directory)
}
\arguments{
\item{file_list}{list of edf files}

\item{varnames}{a vector of variable names from experiment c("TRIALID", "ACCURACY")}

\item{patterns}{a vector of patters to look for ("TRIALID", "!V TRIAL_VAR script")}

\item{output_directory}{where you want files to be stored}

\item{file_list}{list of edf files}

\item{varnames}{a vector of variable names from experiment c("TRIALID", "ACCURACY")}

\item{patterns}{a vector of patters to look for ("TRIALID", "!V TRIAL_VAR script")}

\item{file_list}{list of edf files}

\item{varnames}{a vector of variable names from experiment c("TRIALID", "ACCURACY")}

\item{patterns}{a vector of patters to look for ("TRIALID", "!V TRIAL_VAR script")}
}
\description{
Take EDF files and return a tiblle with important behavioral events

Take EDF files and return a tiblle with important behavioral events
}
