% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpolate_pupil.R
\name{interpolate_pupil}
\alias{interpolate_pupil}
\title{Turns pupil sizes with blinks NA and linearly interpolates values}
\usage{
interpolate_pupil(datafile, extendblinks = FALSE)
}
\arguments{
\item{datafile}{data frame.}

\item{extendblinks}{blinks already extended  in data frame}
}
\value{
data frame containing interpolated data
}
\description{
Turns pupil sizes with blinks NA and linearly interpolates values
}
