% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lowpass_filter.R
\name{lowpass_filter}
\alias{lowpass_filter}
\title{This function performs low pass filtering}
\usage{
lowpass_filter(pd, order, filter_to_use, samplerate = NA)
}
\arguments{
\item{pd}{pupil column.}

\item{order}{order butter filter}

\item{filter_to_use}{hz to low pass filter}

\item{samplerate}{sampling rate of the eye tracker in Hz}
}
\value{
vector containing low pass filtered pupil data
}
\description{
This function performs low pass filtering
}
