% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_asc_files.R
\name{merge_asc_files}
\alias{merge_asc_files}
\title{Take asc files and extract relevant data
Does not merge data with messages for fixation data, obtains sample report and puts time in ms
This code was written by Dr. Holger Mitterer}
\usage{
merge_asc_files(dirList, homeDir = "./")
}
\arguments{
\item{dirList}{list if asc files}

\item{homeDir}{directory of edf files}
}
\description{
Take asc files and extract relevant data
Does not merge data with messages for fixation data, obtains sample report and puts time in ms
This code was written by Dr. Holger Mitterer
}
