% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_asc.R
\name{parse_asc}
\alias{parse_asc}
\title{Take asc files and extract relevant data
Does not merge data with messages for fixation data, obtains sample report and puts time in ms
This code was written by Dr. Holger Mitterer}
\usage{
parse_asc(dirList, homeDir = "./", overwriteBlinks = FALSE, cutPreview = 0)
}
\arguments{
\item{dirList}{list if asc files}

\item{homeDir}{directory of edf files}

\item{overwriteBlinks}{set to false}

\item{cutPreview}{number of ms to cut from the start of each trial (default 0)}
}
\description{
Take asc files and extract relevant data
Does not merge data with messages for fixation data, obtains sample report and puts time in ms
This code was written by Dr. Holger Mitterer
}
