% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess_pupil.R, R/preprocessing_pupil.R
\name{preprocess_pupil}
\alias{preprocess_pupil}
\title{Preprocessing clean-up of messy pupil data.}
\usage{
preprocess_pupil(datafile, baseline_window = NA, binsize = NA,
  stim_onset = NA, smoothing = NA)

preprocess_pupil(datafile, baseline_window = NA, binsize = NA,
  stim_onset = NA, smoothing = NA)
}
\arguments{
\item{datafile}{raw pupil data}

\item{baseline_window}{user-specified threshold for baseline window.}

\item{binsize}{user-specified threshold for binning data.}

\item{stim_onset}{user-specified threshold for when stimulus of interest appears.}

\item{smoothing}{Low-pass filtering or moving average}

\item{datafile}{raw pupil data}

\item{baseline_window}{user-specified threshold for baseline window.}

\item{binsize}{user-specified threshold for binning data.}

\item{stim_onset}{user-specified threshold for when stimulus of interest appears.}

\item{smoothing}{Low-pass filtering or moving average}
}
\value{
data frame containing pre-processed pupil data

data frame containing pre-processed pupil data
}
\description{
Turn all blinks NA, linear interpolation of NA values
user-specified thresholds for baseline values, binsizes, and stimulus onset

Turn all blinks NA, linear interpolation of NA values
user-specified thresholds for baseline values, binsizes, and stimulus onset
}
