% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{WordLearnEx}
\alias{WordLearnEx}
\title{Effect of transitional probability on novel word learning}
\format{A data frame with 560 rows and 4 variables
\describe{
\item{Subject}{subject id, N=56}
\item{TP}{transitional probability of words, manipulated between-subjects}
\item{Block}{training block (1-10), each block consists of 6 trials}
\item{Accuracy}{proportion correct responses during the block, chance = 0.5}
}}
\source{
Partial data from Mirman et al. (2008, Cognition, 108:1, 271-280). \url{http://www.danmirman.org/pdfs/MirmanMagnusonGrafEstesDixon2008.pdf}
}
\usage{
WordLearnEx
}
\description{
Accuracy in 2-alternative forced-choice word learning experiment. On each trial, participants see two complex geometric shapes, hear a novel word (e.g., "pibu"), and have to guess which object corresponds to the word. They receive feedback and gradually learn which object goes with which word. There are 6 trials in each of the 10 training blocks. Transitional probability of the novel words (i.e., p("bu"|"pi")) was experimentally manipulated (between-subjects) during an exposure phase immediately preceding the word learning phase.
}
\keyword{datasets}
